KTO-1037: Refactor "Koulutuksen aloitusajankohta" fields to a separate component
- Use new KoulutuksenAloitusajankohtaFields in HakuForm
